back-end

+++ Student Enroll Management/UI-for-Project/MainWindow.xaml.cs
@@ -12,6 +12,10 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using System.Data.SqlClient;
+using System.Data;
+using UI_for_Project.Controller;
+using UI_for_Project.Model;
 
 namespace UI_for_Project
 {
@@ -20,21 +24,61 @@ namespace UI_for_Project
     /// </summary>
     public partial class MainWindow : Window
     {
+
+
         public MainWindow()
         {
             InitializeComponent();
+            // GIAO DIỆN STATISTIC CHỈ HIỂN THỊ KHI HOÀN THÀNH CHẤM ĐIỂM
+            // NẾU CHƯA HOÀN THÀNH THÌ KHI CLICK VÀO SẼ RA THÔNG BÁO VÀ CHUYỂN VỀ GIAO DIỆN CANDIDATE
+        }
+        private void LogOut(object sender, RoutedEventArgs e)
+        {
+            // click button log_out
+            confirmPerson.setPersonPermission("NONE");
 
             var loginWindow = new Login();
             loginWindow.Show();
             this.Close();
 
         }
+        private void Window_Loaded(object sender, RoutedEventArgs e)
+        {
+            //this.Show();
+
+            // KIỂM TRA THỬ PHẦN ĐĂNG KÍ ĐÃ HOÀN THÀNH CHƯA, NẾU HOÀN THÀNH RỒI THÌ HIỂN THỊ MAIN, NẾU CHƯA THÌ HIỂN THỊ PHẦN ĐĂNG KÍ 
+            if (confirmFinishRegister.getConfirm() == true)
+            {
+                if (confirmPerson.getPersonPermission() == "NONE")
+                {
+                    var intent_login = new Login();
+                    intent_login.Show();
+                    this.Close();
+                }
+                if (confirmPerson.getPersonPermission() == "ADMIN")
+                {
+                    btnLogOut.Content = "log out with admin";
+                }
+                else if (confirmPerson.getPersonPermission() == "USER")
+                {
+                    btnLogOut.Content = "log out with user";
+                }
+            }
+            else
+            {
+                var intent_login = new Login();
+                intent_login.Show();
+                this.Close();
+            }
+        }
+
+
         // functional menu button events
         private void btnCandidate_Click(object sender, RoutedEventArgs e)
         {
             btnCandidate.Style = SelectingStyle();
 
-            if(!(Holder.Content is CandidateList))
+            if (!(Holder.Content is CandidateList))
             {
                 Holder.Content = null;
                 while (Holder.NavigationService.RemoveBackEntry() != null) ;
@@ -57,36 +101,47 @@ namespace UI_for_Project
             {
                 Holder.Content = null;
                 while (Holder.NavigationService.RemoveBackEntry() != null) ;
-                Holder.Content = new Scoring();
+                Holder.Content = new Scoring(OpenScoreEditingUI);
             }
 
             btnCandidate.Style = IdlingStyle();
             btnStatistic.Style = IdlingStyle();
-
-            // model manipulation
         }
 
-        private void btnStatistic_Click(object sender,RoutedEventArgs e)
-        {   
-            btnStatistic.Style = SelectingStyle();
-
-            // remove current content on the page
-            Holder.Content = null;
-            while (Holder.NavigationService.RemoveBackEntry() != null) ;
-
-            // style changing
-            btnCandidate.Style = IdlingStyle();
-            btnScoring.Style = IdlingStyle();
-
-            // model manipulation
-        }
-
-        // Log Out event
-        private void LogOut(object sender, RoutedEventArgs e)
+        private void btnStatistic_Click(object sender, RoutedEventArgs e)
         {
+            if (confirmEditScoring.confirmAllSubject() == true)
+            {
+                btnStatistic.Style = SelectingStyle();
 
+               // remove current content on the page
+            if (!(Holder.Content is Statistic))
+                {
+                    Holder.Content = null;
+                    while (Holder.NavigationService.RemoveBackEntry() != null) ;
+                    Holder.Content = new Statistic();
+                }
+
+               // style changing
+            btnCandidate.Style = IdlingStyle();
+                btnScoring.Style = IdlingStyle();
+            }
+            else
+            {
+                MessageBox.Show("You need finish Scoring before.");
+                btnCandidate.Style = SelectingStyle();
+
+                if (!(Holder.Content is CandidateList))
+                {
+                    Holder.Content = null;
+                    while (Holder.NavigationService.RemoveBackEntry() != null) ;
+                    Holder.Content = new CandidateList();
+                }
+
+                btnScoring.Style = IdlingStyle();
+                btnStatistic.Style = IdlingStyle();
+            }
         }
-
         // changing Style for EveryButton
         Style SelectingStyle()
         {
@@ -98,6 +153,10 @@ namespace UI_for_Project
             return FindResource("FunctionBtn_Idle") as Style;
         }
 
-        
+        void OpenScoreEditingUI(string subjectBtn)
+        {
+            EditScoring a = new EditScoring(subjectBtn);
+            Holder.Content = a;
+        }
     }
 }
